test(utils): move to individual files and migrate to TS
