+++ src/lib/utils/__tests__/capitalize-test.ts
@@ -1,0 +1,7 @@
+import capitalize from '../capitalize';
+
+describe('capitalize', () => {
+  it('should capitalize the first character only', () => {
+    expect(capitalize('hello')).toBe('Hello');
+  });
+});
