+++ src/lib/utils/escape-highlight.ts
@@ -45,7 +45,9 @@ function recursiveEscape(input: any): any {
   };
 }
 
-export function escapeHits<THit extends Hit>(hits: THit[]): EscapedHits<THit> {
+export function escapeHits<THit extends Hit>(
+  hits: THit[] | EscapedHits<THit>
+): EscapedHits<THit> {
   if ((hits as any).__escaped === undefined) {
     // We don't override the value on hit because it will mutate the raw results
     // instead we make a shallow copy and we assign the escaped values on it.
